+++ projects/ngx-simple-gallery/src/public-api.ts
@@ -6,4 +6,4 @@ export * from './lib/core/model/gallery-item';
 export * from './lib/core/model/gallery-config';
 export * from './lib/core/service/gallery.service';
 export * from './lib/ngx-simple-gallery.component';
-export * from './lib/open-gallery/open-gallery.directive';
+export * from './lib/simple-gallery/simple-gallery.directive';